+++ specs/consumer/component/consumer-component.spec.js
@@ -28,7 +28,7 @@ describe('ConsumerComponent', () => {
       sinon.stub(Impl, ['load']).returns('impl');
     });
     it('should throw an error for a mismatch compiler interface', () => {
-      const scope = { loadEnvironment: () => { return {} } };
+      const scope = { loadEnvironment: () => { return {}; } };
       const result = consumerComponent.build({ scope });
       expect(result).to.be.a('Promise');
       return result
@@ -36,7 +36,7 @@ describe('ConsumerComponent', () => {
           throw new Error('Promise should fail');
         })
         .catch((err) => {
-          expect(err).to.eql('"scope/box/name::2" does not have a valid compiler interface');
+          expect(err).to.eql('"scope/box/name::2" does not have a valid compiler interface, it has to return a build method');
         });
     });
     it('should NOT throw an error for a correct compiler interface', () => {