+++ src/npm-client/index.js
@@ -1,0 +1,3 @@
+import npmClient from './npm-client';
+
+export default npmClient;